@MuriloAlmeid subindo configuração do swagger, adicionando depedencia de dto
