+++ src/main/java/sptech/elderly/web/dto/usuario/atualizarEmail.java
@@ -1,4 +1,4 @@
-package sptech.elderly.web.dto;
+package sptech.elderly.web.dto.usuario;
 
 import jakarta.validation.constraints.NotBlank;
