+++ src/main/resources/application.properties
@@ -7,9 +7,19 @@ spring.datasource.url=${DB_URL}
 spring.datasource.username=${DB_USER}
 spring.datasource.password=${DB_PASSWORD}
 
-#JPA configuração para exibir comandos sql
+# JPA configuração para exibir comandos sql
 spring.jpa.show-sql=true
 
-#JPA configuração para formatar o script sql
+# JPA configuração para formatar o script sql
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.hibernate.ddl-auto=validate
+
+# SpringDoc OpenApi 3.1 & Swagger 3 # Configurações do Swagger
+# Habilita o Swagger
+swagger.enabled=true
+
+# Termos de serviço da documentação
+swagger.api-info.terms-of-service-url=https://example.com/terms-of-service
+
+# Pacote base dos controllers para escanear
+swagger.base-package=sptech.elderly.web.controller
